+++ cx-portal/src/components/shared/templates/StaticTemplateResponsive/index.tsx
@@ -18,7 +18,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-import { ProviderProps } from './StaticTypes'
+import { ProviderProps, TemplateNames } from './StaticTypes'
 import TextImageSideBySide from './Cards/TextImageSideBySide'
 import TextVideoSideBySide from './Cards/TextVideoSideBySide'
 import TextImageCenterAligned from './Cards/TextImageCenterAligned'
@@ -29,6 +29,8 @@ import VideoTextSideBySide from './Cards/VideoTextSideBySide'
 import ImageTextCenterAligned from './Cards/ImageTextCenterAligned'
 import TextCenterAlignedBody2 from './Cards/TextCenterAlignedBody2'
 import './StaticTemplate.scss'
+import FlexImages from './Cards/FlexImages'
+import GridImages from './Cards/GridImages'
 
 const TemplateConfig = ({
   provider,
@@ -39,28 +41,28 @@ const TemplateConfig = ({
 }) => {
   switch (provider.template) {
     //Text in the left and image in the right side
-    case 'TextImageSideBySide':
+    case TemplateNames.TextImageSideBySide:
       return <TextImageSideBySide baseUrl={baseUrl} provider={provider} />
     //Text in the left and video in the right side
-    case 'TextVideoSideBySide':
+    case TemplateNames.TextVideoSideBySide:
       return <TextVideoSideBySide provider={provider} />
     //video in the left and Text in the right side
-    case 'VideoTextSideBySide':
+    case TemplateNames.VideoTextSideBySide:
       return <VideoTextSideBySide provider={provider} />
     //Text and image component both center aligned
-    case 'TextImageCenterAligned':
+    case TemplateNames.TextImageCenterAligned:
       return <TextImageCenterAligned baseUrl={baseUrl} provider={provider} />
     //Image and text component both center aligned
-    case 'ImageTextCenterAligned':
+    case TemplateNames.ImageTextCenterAligned:
       return <ImageTextCenterAligned baseUrl={baseUrl} provider={provider} />
     //Text component center aligned
-    case 'TextCenterAligned':
+    case TemplateNames.TextCenterAligned:
       return <TextCenterAligned provider={provider} />
     //Text component body2
-    case 'TextCenterAlignedBody2':
+    case TemplateNames.TextCenterAlignedBody2:
       return <TextCenterAlignedBody2 provider={provider} />
     //Combination of Text and image component both center aligned with grid layout card component
-    case 'TextImageCenterAlignedWithCardGrid':
+    case TemplateNames.TextImageCenterAlignedWithCardGrid:
       return (
         <>
           <TextImageCenterAligned baseUrl={baseUrl} provider={provider} />
@@ -73,7 +75,7 @@ const TemplateConfig = ({
         </>
       )
     //Combination of Text center aligned component with grid layout card component
-    case 'TextCenterAlignedWithCardGrid':
+    case TemplateNames.TextCenterAlignedWithCardGrid:
       return (
         <>
           <TextCenterAligned provider={provider} />
@@ -86,7 +88,7 @@ const TemplateConfig = ({
         </>
       )
     //Combination of Text and image side by side with grid layout card component
-    case 'TextImageSideBySideWithCardGrid':
+    case TemplateNames.TextImageSideBySideWithCardGrid:
       return (
         <>
           <TextImageSideBySide baseUrl={baseUrl} provider={provider} />
@@ -98,11 +100,31 @@ const TemplateConfig = ({
         </>
       )
     //Combination of Text center aligned component with grid layout link component
-    case 'TextCenterAlignedWithLinkButtonGrid':
+    case TemplateNames.TextCenterAlignedWithLinkButtonGrid:
       return (
         <>
           <TextCenterAligned provider={provider} />
           <LinkButtonGrid provider={provider} grid={provider.grid} />
+        </>
+      )
+    //Combination of Text center aligned component with images in a row
+    case TemplateNames.TextCenterAlignedWithImagesInFlex:
+      return (
+        <>
+          <TextCenterAligned provider={provider} />
+          <FlexImages provider={provider} baseUrl={baseUrl} />
+        </>
+      )
+    //Combination of Text center aligned component with images in grid fashion
+    case TemplateNames.TextCenterAlignedWithImagesInGrid:
+      return (
+        <>
+          <TextCenterAligned provider={provider} />
+          <GridImages
+            provider={provider}
+            baseUrl={baseUrl}
+            grid={provider.grid}
+          />
         </>
       )
     default: